refactor: change direction breadCrumb function
